backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
apscheduler==3.10.4

# ML / Data Science
tensorflow>=2.15.0
xgboost>=2.0.0
lightgbm>=4.0.0
scikit-learn>=1.4.0
joblib>=1.3.0

# Data
pandas>=2.1.0
numpy>=1.26.0
yfinance>=0.2.30
fredapi>=0.5.1
newsapi-python>=0.2.7
vaderSentiment>=3.3.2
requests>=2.31.0
